HTBHF-540 move logger to avoid circular dependencies
