+++ src/web/server/error-handlers.js
@@ -1,5 +1,5 @@
 const httpStatus = require('http-status-codes')
-const { logger } = require('./logger')
+const { logger } = require('../logger')
 
 const logErrors = (err, req, res, next) => {
   logger(req).log({
